requirement.txt
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.2
opencv-contrib-python==4.9.0.80
pillow==10.2.0
streamlit